+++ lib/romp_crm/accounts/user.ex
@@ -32,6 +32,8 @@ defmodule RompCrm.Accounts.User do
     field :sms_reminders_enabled, :boolean, default: false
     field :sms_reminder_prefs_json, :string
 
+    field :sms_assistant_intro_completed_at, :utc_datetime
+
     has_many :business_memberships, RompCrm.Businesses.BusinessMembership
 
     timestamps(type: :utc_datetime)